% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{function1}
\alias{function1}
\title{Function1}
\usage{
function1(x)
}
\arguments{
\item{x}{input data frame}
}
\description{
Clean data frame
}
\examples{
function1(data.frame(col1 = c(1, 2), col2 = c(3, 4)))
}
