% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{function2}
\alias{function2}
\title{Show output for y}
\usage{
function2(y)
}
\arguments{
\item{y}{input y}
}
\description{
Show output for y
}
\examples{
function2(data.frame(col1 = c(1, 2), col2 = c(3, 4)))
}
