% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPipes-package.R
\docType{package}
\name{testPipes-package}
\alias{testPipes}
\alias{testPipes-package}
\title{testPipes: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/janclod/testPipes}
  \item Report bugs at \url{https://github.com/janclod/testPipes/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
